Move test method to test class.
